-added var. -support preincrement ++ -optimized code to asm

+++ src/main.cpp
@@ -9,35 +9,40 @@ int main(int argc, char * argv[]){
 	int i=0;
 	i=10*-2;
 
-
+	print_info_cr("sizeof(CObject)=%i sizeof(float)=%i sizeof(string)=%i",sizeof(CObject),sizeof(float),sizeof(string));
 
 	i=i---i;
 
 	if(argc < 2){
-		printf("Put expression to evaluate.\n");
+		printf("Put file to evaluate.\n");
 		printf("\n");
 		printf("Example:\n");
 		printf("\n");
-		printf("5*6+3");
+		printf("file.zs");
 		printf("\n");
 		printf("\n");
 		return 0;
 	}
-	
-	CZG_Script *zg_script = CZG_Script::getInstance();
-	zg_script->init();
+	ByteBuffer *buffer = CIO_Utils::readFile(argv[1]);
 
-	zg_script->eval(argv[1]);
+	if(buffer != NULL){
 
-	print_info_cr("sizeobject:%i",sizeof(CObject));
-	print_info_cr("sizenumber:%i",sizeof(CNumber));
+		CZG_Script *zg_script = CZG_Script::getInstance();
+		zg_script->init();
 
-	print_info_cr("float:%s",typeid(float).name());
-	print_info_cr("string:%s",typeid(string).name());
-	print_info_cr("bool:%s",typeid(bool).name());
+		zg_script->eval((char *)buffer->data_buffer);
+
+		print_info_cr("sizeobject:%i",sizeof(CObject));
+		print_info_cr("sizenumber:%i",sizeof(CNumber));
+
+		print_info_cr("float:%s",typeid(float).name());
+		print_info_cr("string:%s",typeid(string).name());
+		print_info_cr("bool:%s",typeid(bool).name());
 
 
-	CZG_Script::destroy();
+		CZG_Script::destroy();
+		delete buffer;
+	}
 
 #if defined(__DEBUG__) && defined(__MEMMANAGER__)
   MEM_ViewStatus();
